Major updates for Ubuntu 18.04
This was a major update in preparation for Ubuntu 18.04.

The ansible/ directory was flattened and the patches were finally
removed after removing the FGtatsuro roles. The virtualbox role will
have to be added back later once a new role is found.

The install software shell no longer executes the ansible commands but
does perform a privilege escalation if the file is run without sudo.